+++ PKHeX.Core/Legality/Moves/Breeding/MoveBreed6.cs
@@ -14,7 +14,7 @@ public static class MoveBreed6
     private const int level = 1;
 
     /// <inheritdoc cref="MoveBreed.Validate"/>
-    public static bool Validate(int generation, ushort species, int form, GameVersion version, ReadOnlySpan<ushort> moves, Span<byte> origins)
+    public static bool Validate(int generation, ushort species, byte form, GameVersion version, ReadOnlySpan<ushort> moves, Span<byte> origins)
     {
         var count = moves.IndexOf((ushort)0);
         if (count == 0)
