+++ PKHeX.Core/Legality/Verifiers/ParseSettings.cs
@@ -41,7 +41,7 @@ public static class ParseSettings
     public static IReadOnlyList<string> MoveStrings { get; private set; } = Util.GetMovesList(GameLanguage.DefaultLanguage);
     public static IReadOnlyList<string> SpeciesStrings { get; private set; } = Util.GetSpeciesList(GameLanguage.DefaultLanguage);
     public static string GetMoveName(ushort move) => move >= MoveStrings.Count ? LegalityCheckStrings.L_AError : MoveStrings[move];
-    public static IEnumerable<string> GetMoveNames(IEnumerable<ushort> moves) => moves.Select(m => (uint)m >= MoveStrings.Count ? LegalityCheckStrings.L_AError : MoveStrings[m]);
+    public static IEnumerable<string> GetMoveNames(IEnumerable<ushort> moves) => moves.Select(m => m >= MoveStrings.Count ? LegalityCheckStrings.L_AError : MoveStrings[m]);
 
     public static void ChangeLocalizationStrings(IReadOnlyList<string> moves, IReadOnlyList<string> species)
     {
